requirements.txt
kafka-python==2.0.2
pandas==1.4.1
requests==2.27.1
